[#383] Move current case page into old-cases
