+++ frontend/src/pages/old-cases/cases.scss
