+++ frontend/src/pages/old-cases/components/AreaUnitFields.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { Form, Select, Row, Col } from "antd";
-import { selectProps } from "./";
+import { selectProps } from ".";
 import { areaUnitOptions, volumeUnitOptions } from "../../../store/static";
 
 const responsiveCol = {
