+++ frontend/src/pages/old-cases/components/DashboardSensitivityAnalysis.js
