+++ frontend/src/pages/old-cases/visualizations/ChartBinningHeatmap.js
