+++ frontend/src/pages/old-cases/visualizations/ChartCurrentFeasible.js
