+++ frontend/src/pages/old-cases/visualizations/ChartIncomeGap.js
