+++ frontend/src/pages/old-cases/visualizations/ChartIncomeLevelPerCommodities.js
