+++ frontend/src/pages/old-cases/visualizations/ChartSensitivityAnalysisLine.js
