+++ frontend/src/pages/old-cases/visualizations/SegmentSelector.js
